+++ attributes/default.rb
@@ -1,7 +1,17 @@
 # CRAN
-default['rstudio']['cran']['uri'] = 'http://cran.stat.ucla.edu/bin/linux/ubuntu'
-default['rstudio']['cran']['key'] = 'E084DAB9'
-default['rstudio']['cran']['keyserver'] = 'keyserver.ubuntu.com'
+default['rstudio']['cran']['mirror'] = 'http://cran.rstudio.com/'
+
+# APT configuration for Ubuntu or Debian installs.
+case node["platform"].downcase  
+when "ubuntu"
+    default['rstudio']['apt']['key'] = 'E084DAB9'
+    default['rstudio']['apt']['keyserver'] = 'keyserver.ubuntu.com'
+    default['rstudio']['apt']['uri'] = 'http://cran.stat.ucla.edu/bin/linux/ubuntu'
+when "debian"
+    default['rstudio']['apt']['key'] = '381BA480'
+    default['rstudio']['apt']['keyserver'] = 'subkeys.pgp.net'
+    default['rstudio']['apt']['uri'] = 'http://cran.stat.ucla.edu/bin/linux/debian'
+end
 
 # You can define a simple array of packages in your role/environment/node and the 
 # CRAN recipe will install them.
@@ -23,7 +33,20 @@ default['rstudio']['session']['cran_repo'] = 'http://cran.case.edu/'
 default['rstudio']['nginx']['port'] = '80'
 default['rstudio']['nginx']['server_name'] = ''
 default['rstudio']['nginx']['location'] = '/'
+default['rstudio']['nginx']['shiny_location'] = ''
 
 # SSL certificates
 default['rstudio']['ssl']['crt_file'] = ''
 default['rstudio']['ssl']['key_file'] = ''
+
+# Shiny server
+default['rstudio']['shiny']['user'] = 'shiny'
+default['rstudio']['shiny']['www_port'] = '8100'
+default['rstudio']['shiny']['www_address'] = '127.0.0.1'
+default['rstudio']['shiny']['site_dir'] = '/srv/shiny-server'
+default['rstudio']['shiny']['log_dir'] = '/var/log/shiny-server'
+default['rstudio']['shiny']['directory_index'] = 'on'
+
+# Shiny server supports the users cookbook for HTTP Auth
+default['rstudio']['shiny']['htpasswd_file'] = ''
+default['rstudio']['shiny']['htpasswd_group'] = 'sysadmin'