+++ metadata.rb
@@ -4,7 +4,8 @@ maintainer_email 'joe@joestump.net'
 license          'All rights reserved'
 description      'Installs/Configures rstudio'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.0'
+version          '0.2.0'
 
 depends "apt"
 depends "nginx"
+depends "r"