+++ CHANGELOG.md
@@ -4,3 +4,12 @@ rstudio CHANGELOG
 0.1.0
 -----
 - Initial release of rstudio (Joe Stump)
+
+0.2.0
+-----
+- Added [Shiny Server](http://www.rstudio.com/shiny/server/) support (Joe Stump)
+- Added tentative support for Debian (Joe Stump)
+- Updated the APT installation to use the official APT repositories for Ubuntu and Debian (Joe Stump)
+- Updated the Nginx configuration to support serving both RStudio and Shiny from the same server (Joe Stump)
+- Now depends on the [r](https://github.com/stevendanna/cookbook-r/) cookbook.
+- Fixed support for `node['rstudio']['cran']['packages']` using `r_package` LWRP from the `r` cookbook.